+++ src/plone/restapi/serializer/catalog.py
@@ -2,11 +2,19 @@
 from plone.restapi.batching import HypermediaBatch
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.interfaces import ISerializeToJsonSummary
-from Products.ZCatalog.Lazy import Lazy
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.interface import implementer
 from zope.interface import Interface
+
+import logging
+
+try:
+    from ZTUtils.Lazy import Lazy
+except ImportError:
+    from Products.ZCatalog.Lazy import Lazy
+
+log = logging.getLogger(__name__)
 
 
 @implementer(ISerializeToJson)
@@ -33,9 +41,19 @@ class LazyCatalogResultSerializer(object):
         results["items"] = []
         for brain in batch:
             if fullobjects:
-                result = getMultiAdapter(
-                    (brain.getObject(), self.request), ISerializeToJson
-                )(include_items=False)
+                try:
+                    result = getMultiAdapter(
+                        (brain.getObject(), self.request), ISerializeToJson
+                    )(include_items=False)
+                except KeyError:
+                    # Guard in case the brain returned refers to an object that doesn't
+                    # exists because it failed to uncatalog itself or the catalog has
+                    # stale cataloged objects for some reason
+                    log.warning(
+                        "Brain getObject error: {} doesn't exist anymore".format(
+                            brain.getPath()
+                        )
+                    )
             else:
                 result = getMultiAdapter(
                     (brain, self.request), ISerializeToJsonSummary