+++ src/plone/restapi/tests/test_comments.py
@@ -10,6 +10,7 @@ from unittest import TestCase
 from zope.component import createObject
 from zope.component import getMultiAdapter
 from zope.component import getUtility
+from zope.component import queryUtility
 
 
 class TestCommentsSerializers(TestCase):
@@ -86,3 +87,69 @@ class TestCommentsSerializers(TestCase):
         self.assertEqual(set(output), set(expected))
 
         self.assertEqual(set(output["text"]), set(["data", "mime-type"]))
+
+    def test_comment_with_mimetype_text_plain(self):
+
+        self.conversation = IConversation(self.doc)
+        self.replies = IReplies(self.conversation)
+        comment = createObject("plone.Comment")
+        comment.text = "Hey, I am plain text!"
+        comment.mime_type = "text/plain"
+        self.comment = self.replies[self.replies.addComment(comment)]
+
+        serializer = getMultiAdapter((self.comment, self.request), ISerializeToJson)
+
+        # serializer should return HTML with a clickable link
+        self.assertEqual(
+            'Hey, I am plain text!',
+            serializer()["text"]["data"],
+        )
+        # serializer should return mimetype = text/x-web-intelligent
+        self.assertEqual("text/plain", serializer()["text"]["mime-type"])
+
+    def test_comment_with_mimetype_intelligent_text(self):
+        # Set text transform to intelligent text
+        registry = queryUtility(IRegistry)
+        settings = registry.forInterface(IDiscussionSettings, check=False)
+        settings.text_transform = "text/x-web-intelligent"
+
+        self.conversation = IConversation(self.doc)
+        self.replies = IReplies(self.conversation)
+        comment = createObject("plone.Comment")
+        comment.text = "Go to https://www.plone.org"
+        comment.mime_type = "text/x-web-intelligent"
+        self.comment = self.replies[self.replies.addComment(comment)]
+
+        serializer = getMultiAdapter((self.comment, self.request), ISerializeToJson)
+
+        # serializer should return HTML with a clickable link
+        self.assertEqual(
+            'Go to <a href="https://www.plone.org" '
+            + 'rel="nofollow">https://www.plone.org</a>',
+            serializer()["text"]["data"],
+        )
+        # serializer should return mimetype = text/html
+        self.assertEqual("text/html", serializer()["text"]["mime-type"])
+
+    def test_comment_with_mimetype_html(self):
+        # Set text transform to text/html
+        registry = queryUtility(IRegistry)
+        settings = registry.forInterface(IDiscussionSettings, check=False)
+        settings.text_transform = "text/html"
+
+        self.conversation = IConversation(self.doc)
+        self.replies = IReplies(self.conversation)
+        comment = createObject("plone.Comment")
+        comment.text = "Go to <a href='https://www.plone.org'>Plone</a>"
+        comment.mime_type = "text/html"
+        self.comment = self.replies[self.replies.addComment(comment)]
+
+        serializer = getMultiAdapter((self.comment, self.request), ISerializeToJson)
+
+        # serializer should return HTML
+        self.assertEqual(
+            'Go to <a href="https://www.plone.org">Plone</a>',
+            serializer()["text"]["data"],
+        )
+        # serializer should return mimetype = text/html
+        self.assertEqual("text/html", serializer()["text"]["mime-type"])